fixing type translation

+++ src/main/java/com/akiban/sql/pg/PostgresType.java
@@ -382,7 +382,7 @@ public class PostgresType extends ServerType
             }
             else {
                 akType = AkType.LONG;
-                tInstance = MNumeric.BIGINT.instance();
+                tInstance = MNumeric.INT.instance();
             }
             break;
         case TypeId.FormatIds.LONGINT_TYPE_ID:
